Fix priority field type mismatch with MongoDB validator
Change Priority field from int to int64 to match MongoDB BSON type:
- Update BusinessUnit.Priority from int to int64
- Update BusinessUnitUpdate.Priority from *int to *int64
- Update NormalizePriority function signature to use int64

Prevents data corruption on 32-bit systems where int is 32 bits
but MongoDB validator expects 64-bit long type.

Fixes: Code Review Issue #7

+++ pkg/model/business_unit.go
@@ -9,7 +9,7 @@ type BusinessUnit struct {
 	Labels      []string  `bson:"labels" validate:"required,min=1,max=10,dive,required"`
 	AdminPhone  string    `bson:"admin_phone" validate:"required,e164,supported_country"`
 	Maintainers []string  `bson:"maintainers" validate:"omitempty,dive,required"`
-	Priority    int       `bson:"priority" validate:"omitempty,min=0"`
+	Priority    int64     `bson:"priority" validate:"omitempty,min=0"`
 	TimeZone    string    `bson:"time_zone" validate:"omitempty,timezone"`
 	WebsiteURL  string    `json:"website_url" bson:"website_url,omitempty" validate:"omitempty,url,startswith=https://"`
 	CreatedAt   time.Time `bson:"created_at" validate:"omitempty"`
@@ -21,7 +21,7 @@ type BusinessUnitUpdate struct {
 	Labels      []string  `json:"labels,omitempty" validate:"omitempty,min=1,max=10,dive,required"`
 	AdminPhone  string    `json:"admin_phone,omitempty" validate:"omitempty,e164,supported_country"`
 	Maintainers *[]string `json:"maintainers,omitempty" validate:"omitempty,dive,required"`
-	Priority    *int      `json:"priority,omitempty" validate:"omitempty,min=0"`
+	Priority    *int64    `json:"priority,omitempty" validate:"omitempty,min=0"`
 	TimeZone    string    `json:"time_zone,omitempty" validate:"omitempty,timezone"`
 	WebsiteURL  *string   `json:"website_url,omitempty" validate:"omitempty,url,startswith=https://"`
 }
